pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.9"

dependencies = [
    # LiveKit Agents (pinned for stability)
    "livekit-agents[deepgram,silero,turn-detector]==1.2.0",
    "livekit-plugins-noise-cancellation==0.2.3",
    "livekit-plugins-elevenlabs==1.0.14",
    
    # Database & ORM
    "sqlalchemy[asyncio]==2.0.23",
    "asyncpg==0.29.0",
    "alembic>=1.13",
    
    # Caching & State
    "redis[hiredis]==5.0.1",
    
    # Resilience & Observability
    "tenacity==8.2.3",
    "structlog==24.1.0",
    
    # Configuration
    "python-dotenv==1.0.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
